refactor: rename folder and remove afmLocale
