+++ src/lib/components/CustomAMap/index.tsx
@@ -4,7 +4,7 @@ import { Map, Marker, MapProps } from "react-amap";
 import Geolocation from "react-amap-plugin-custom-geolocation";
 import CurrentAddress from "./CurrentAddress";
 import PlaceSearch from "./PlaceSearch";
-import { ConfigContext } from '../../../config-provider/context';
+import { ConfigContext } from '../../../config-context/context';
 import { Position } from './Props';
 
 const defaultPlugins = ["Scale"];
